ci: start test-validator in build
