+++ .github/workflows/test.yml
@@ -47,17 +47,8 @@ jobs:
 
             - name: build with validator
               run: |
-                  (
-                    mkdir -p test/.cache/
-                    cd test/.cache/
-                    solana-test-validator --reset > /dev/null &
-                    validator_pid=$!
-                    trap "kill $validator_pid" EXIT
-                    sleep 5
-                    cd ../..
-                    export PATH="$HOME/.local/share/solana/install/active_release/bin:$PATH"
-                    python3 test/build_comptoken_program.py --features testmode
-                  )
+                  export PATH="$HOME/.local/share/solana/install/active_release/bin:$PATH"
+                  python3 test/build_comptoken_program.py --features testmode
 
             - name: Upload build artifacts
               uses: actions/upload-artifact@v4
@@ -140,7 +131,7 @@ jobs:
               run: npm clean-install --prefix ./test/compto-test-client
 
             - name: run integration tests
-              run: python3 test/full_deploy_test.py --no-build --no-reset # calls node.js scripts to run tests; do not reset the validator
+              run: python3 test/full_deploy_test.py --no-build --no-reset -vvv # calls node.js scripts to run tests; do not reset the validator
 
             - name: Validator logs on failure
               if: failure()
